requirements.txt
pandas==2.2.2
duckdb==1.0.0
pandera==0.20.4
python-dateutil==2.9.0.post0
pyyaml==6.0.2
